[SofaCUDA] Move collision components
