+++ applications/plugins/SofaCUDA/sofa/gpu/cuda/CudaSphereModel.h
@@ -21,30 +21,7 @@
 ******************************************************************************/
 #pragma once
 
-#include <sofa/gpu/cuda/CudaTypes.h>
-#include <sofa/component/collision/geometry/SphereModel.h>
+#include <SofaCUDA/config.h>
 
-namespace sofa::gpu::cuda
-{
-
-using CudaSphereCollisionModel = sofa::component::collision::geometry::SphereCollisionModel<CudaVec3Types>;
-using CudaSphereCollisionModelf1 = sofa::component::collision::geometry::SphereCollisionModel<CudaVec3f1Types>;
-
-using CudaSphere = sofa::component::collision::geometry::TSphere<CudaVec3fTypes>;
-
-} // namespace sofa::gpu::cuda
-
-
-namespace sofa::component::collision::geometry
-{
-
-#if !defined(SOFA_GPU_CUDA_CUDASPHEREMODEL_CPP)
-extern template class SOFA_GPU_CUDA_API SphereCollisionModel<sofa::gpu::cuda::CudaVec3fTypes>;
-extern template class SOFA_GPU_CUDA_API SphereCollisionModel<sofa::gpu::cuda::CudaVec3f1Types>;
-#ifdef SOFA_GPU_CUDA_DOUBLE
-extern template class SOFA_GPU_CUDA_API SphereCollisionModel<sofa::gpu::cuda::CudaVec3dTypes>;
-extern template class SOFA_GPU_CUDA_API SphereCollisionModel<sofa::gpu::cuda::CudaVec3d1Types>;
-#endif // SOFA_GPU_CUDA_DOUBLE
-#endif
-
-} // namespace sofa::component::collision
+SOFA_DEPRECATED_HEADER("v23.06", "v23.12", "SofaCUDA/component/collision/geometry/CudaSphereModel.h")
+#include <SofaCUDA/component/collision/geometry/CudaSphereModel.h>
