+++ applications/plugins/SofaCUDA/sofa/gpu/cuda/CudaTriangleModel.h
@@ -21,31 +21,7 @@
 ******************************************************************************/
 #pragma once
 
-#include <sofa/gpu/cuda/CudaTypes.h>
-#include <sofa/component/collision/geometry/TriangleModel.h>
+#include <SofaCUDA/config.h>
 
-namespace sofa::gpu::cuda
-{
-
-SOFA_CUDA_ATTRIBUTE_DEPRECATED("v22.06 (PR #2673)", "CudaTriangleCollisionModel")
-CudaDeprecatedAndRemoved CudaTriangleModel;
-
-using CudaTriangleCollisionModel = sofa::component::collision::geometry::TriangleCollisionModel<CudaVec3Types>;
-using CudaTriangleCollisionModelf1 = sofa::component::collision::geometry::TriangleCollisionModel<CudaVec3f1Types>;
-
-using CudaTriangle = sofa::component::collision::geometry::TTriangle<CudaVec3fTypes>;
-
-} // namespace sofa::gpu::cuda
-
-namespace sofa::component::collision::geometry
-{
-#if !defined(SOFA_GPU_CUDA_CUDATRIANGLEMODEL_CPP)
-extern template class SOFA_GPU_CUDA_API TriangleCollisionModel<sofa::gpu::cuda::CudaVec3fTypes>;
-extern template class SOFA_GPU_CUDA_API TriangleCollisionModel<sofa::gpu::cuda::CudaVec3f1Types>;
-#ifdef SOFA_GPU_CUDA_DOUBLE
-extern template class SOFA_GPU_CUDA_API TriangleCollisionModel<sofa::gpu::cuda::CudaVec3dTypes>;
-extern template class SOFA_GPU_CUDA_API TriangleCollisionModel<sofa::gpu::cuda::CudaVec3d1Types>;
-#endif  // SOFA_GPU_CUDA_DOUBLE
-#endif
-
-}  // namespace sofa::component::collision::geometry
+SOFA_DEPRECATED_HEADER("v23.06", "v23.12", "SofaCUDA/component/collision/geometry/CudaTriangleModel.h")
+#include <SofaCUDA/component/collision/geometry/CudaTriangleModel.h>
