+++ applications/plugins/SofaCUDA/src/SofaCUDA/component/collision/geometry/CudaLineModel.cpp
