+++ applications/plugins/SofaCUDA/src/SofaCUDA/component/collision/geometry/CudaLineModel.h
@@ -21,7 +21,32 @@
 ******************************************************************************/
 #pragma once
 
-#include <SofaCUDA/config.h>
+#include <sofa/gpu/cuda/CudaTypes.h>
+#include <sofa/component/collision/geometry/LineModel.h>
 
-SOFA_DEPRECATED_HEADER("v23.06", "v23.12", "SofaCUDA/component/collision/geometry/CudaLineModel.h")
-#include <SofaCUDA/component/collision/geometry/CudaLineModel.h>
+namespace sofa::gpu::cuda
+{
+
+SOFA_CUDA_ATTRIBUTE_DEPRECATED("v22.06 (PR #2673)", "CudaLineCollisionModel")
+CudaDeprecatedAndRemoved CudaLineModel;
+
+using CudaLineCollisionModel = sofa::component::collision::geometry::LineCollisionModel<CudaVec3Types>;
+using CudaLineCollisionModelf1 = sofa::component::collision::geometry::LineCollisionModel<CudaVec3f1Types>;
+
+using CudaLine = sofa::component::collision::geometry::TLine<CudaVec3fTypes>;
+
+} // namespace sofa::gpu::cuda
+
+
+namespace sofa::component::collision::geometry
+{
+#if !defined(SOFA_GPU_CUDA_CUDALINEMODEL_CPP)
+extern template class SOFA_GPU_CUDA_API LineCollisionModel<sofa::gpu::cuda::CudaVec3fTypes>;
+extern template class SOFA_GPU_CUDA_API LineCollisionModel<sofa::gpu::cuda::CudaVec3f1Types>;
+#ifdef SOFA_GPU_CUDA_DOUBLE
+extern template class SOFA_GPU_CUDA_API LineCollisionModel<sofa::gpu::cuda::CudaVec3dTypes>;
+extern template class SOFA_GPU_CUDA_API LineCollisionModel<sofa::gpu::cuda::CudaVec3d1Types>;
+#endif  // SOFA_GPU_CUDA_DOUBLE
+#endif
+
+}  // namespace sofa::component::collision::geometry
