+++ applications/plugins/SofaCUDA/src/SofaCUDA/component/collision/geometry/CudaSphereModel.cpp
