+++ applications/plugins/SofaCUDA/src/SofaCUDA/component/collision/geometry/CudaSphereModel.h
@@ -21,7 +21,30 @@
 ******************************************************************************/
 #pragma once
 
-#include <SofaCUDA/config.h>
+#include <sofa/gpu/cuda/CudaTypes.h>
+#include <sofa/component/collision/geometry/SphereModel.h>
 
-SOFA_DEPRECATED_HEADER("v23.06", "v23.12", "SofaCUDA/component/collision/geometry/CudaLineModel.h")
-#include <SofaCUDA/component/collision/geometry/CudaLineModel.h>
+namespace sofa::gpu::cuda
+{
+
+using CudaSphereCollisionModel = sofa::component::collision::geometry::SphereCollisionModel<CudaVec3Types>;
+using CudaSphereCollisionModelf1 = sofa::component::collision::geometry::SphereCollisionModel<CudaVec3f1Types>;
+
+using CudaSphere = sofa::component::collision::geometry::TSphere<CudaVec3fTypes>;
+
+} // namespace sofa::gpu::cuda
+
+
+namespace sofa::component::collision::geometry
+{
+
+#if !defined(SOFA_GPU_CUDA_CUDASPHEREMODEL_CPP)
+extern template class SOFA_GPU_CUDA_API SphereCollisionModel<sofa::gpu::cuda::CudaVec3fTypes>;
+extern template class SOFA_GPU_CUDA_API SphereCollisionModel<sofa::gpu::cuda::CudaVec3f1Types>;
+#ifdef SOFA_GPU_CUDA_DOUBLE
+extern template class SOFA_GPU_CUDA_API SphereCollisionModel<sofa::gpu::cuda::CudaVec3dTypes>;
+extern template class SOFA_GPU_CUDA_API SphereCollisionModel<sofa::gpu::cuda::CudaVec3d1Types>;
+#endif // SOFA_GPU_CUDA_DOUBLE
+#endif
+
+} // namespace sofa::component::collision
