+++ applications/plugins/SofaCUDA/src/SofaCUDA/component/collision/geometry/CudaTriangleModel.cpp
